Gui improvements

+++ GsKit/widgets/GsNumberControl.cpp
@@ -249,7 +249,9 @@ void CGUINumberControl::processRender(const GsRect<float> &RectDispCoordFloat)
     else
         lComp = 0xFF - (mLightRatio*(0xFF-0xDF)/255);
 
-    const Uint32 fillColor = blitsfc.mapRGBA( lComp, lComp, lComp, 0xFF);
+    const Uint32 fillColor = blitsfc.mapRGBA( Uint8(lComp),
+                                              Uint8(lComp),
+                                              Uint8(lComp), 0xFF);
 
     GsRect<Uint16> rect(lRect);
